X-NWB: Rename utils.py to conversion_utils.py
